지난 주차 피드백 반영
- 외부에서 호출하지 않는 메서드의 접근제한자를 private으로 변경
- 구간 추가에서 Section 엔티티를 만들어 넘기지 않고, 구간 추가 필요한 값만 넘겨 의존성 최소화
- SectionService를 제거하고 LineService로 합침
